improvements to threading
creating threads now doesn't require the thread id to be given, a thread id var is incremented and given as an optional argument. Arguments for create thread syscall have now changed, registers for each argument are commented in the syscall handler (essentially, you can now give it the function, argument, and an optional argument in r5 for the thread id that was set that can later be used for syscalls relating to threads, r5 MUST BE SET TO 0 if not wanting to make use of the thread id).